confirms delete request

+++ app/Controllers/TaskListController.js
@@ -1,5 +1,6 @@
 import { ProxyState } from "../AppState.js";
 import { taskListService } from "../Services/TaskListService.js";
+import { generateId } from "../Utils/generateId.js";
 
 function _drawTaskList(){  
   let tasks = ProxyState.tasks;
@@ -11,6 +12,7 @@ function _drawTaskList(){
 export class  TaskListController{
 
   constructor(){
+    ProxyState.on('tasks',_drawTaskList)
     ProxyState.on('tasks',_drawTaskList)
     taskListService.getMyTasks()
     _drawTaskList()
@@ -38,9 +40,28 @@ export class  TaskListController{
     form.reset() 
     _drawTaskList() 
 }
-removeTask(id) { 
-  taskListService.removeTask(id)
+  removeTask(id) { 
+//@ts-ignore
+swal({
+  title: "Are you sure?",
+  text: "Once deleted, you will not be able to recover this task!",
+  icon: "warning",
+  buttons: true,
+  dangerMode: true,
+})
+.then((willDelete) => {
+  if (willDelete) {
+    taskListService.removeTask(id)
+//@ts-ignore
+    swal("Poof! Your task has been deleted!", {
+      icon: "success",
+    });
+  } else {
+    swal("Your task is safe!");
+  }
+})
+
+
+
 }
-
-
 }
